IDs changes to Integer User Rest Controller basic CRUD methods added

+++ src/main/java/com/kopec/wojciech/enginners_thesis/model/AbstractEntity.java
@@ -14,7 +14,7 @@ public abstract class AbstractEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Getter @Setter
-    protected long id;
+    protected Integer id;
 
     @Override
     public String toString() {
