+++ src/main/java/com/kopec/wojciech/enginners_thesis/repository/AmenityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path = "amenities", collectionResourceRel = "amenities")
-public interface AmenityRepository extends JpaRepository<Amenity, Long> {
+public interface AmenityRepository extends JpaRepository<Amenity, Integer> {
 }
